+++ praetorian_cli/sdk/test/test_agent.py
@@ -1,0 +1,23 @@
+import pytest
+
+from praetorian_cli.sdk.model.globals import Risk
+from praetorian_cli.sdk.test.utils import make_test_values, clean_test_entities, setup_chariot
+
+
+@pytest.mark.coherence
+class TestRisk:
+
+    def setup_class(self):
+        self.sdk = setup_chariot()
+        make_test_values(self)
+
+    def test_attribution(self):
+        # create the asset and risk to run attribution
+        result = self.sdk.assets.add(self.asset_dns, self.asset_name)
+        assert result['key'] == self.asset_key
+        result = self.sdk.risks.add(self.asset_key, self.risk_name, Risk.TRIAGE_HIGH.value, self.comment)
+        assert result['key'] == self.risk_key
+        result = self.sdk.agent.attribution(self.risk_key)
+
+    def teardown_class(self):
+        clean_test_entities(self.sdk, self)